@@ -1,95 +1,85 @@
 // Run this and then telnet to localhost:2000 and chat with the bot
 
-var net             = require("net");
-var superscript     = require("superscript");
-var mongoose        = require("mongoose");
-var facts           = require("sfacts");
-var factSystem      = facts.create('telnetFacts');
-mongoose.connect('mongodb://localhost/superscriptDB');
+import net from 'net';
+import superscript from 'superscript';
 
-var options = {};
-var sockets = [];
+const sockets = [];
 
-var TopicSystem = require("superscript/lib/topics/index")(mongoose, factSystem);
-
-options['factSystem'] = factSystem;
-options['mongoose'] = mongoose;
-
-var botHandle = function(err, bot) {
-    
-  var receiveData = function(socket, bot, data) {
+const botHandle = function botHandle(err, bot) {
+  const receiveData = function receiveData(socket, bot, data) {
     // Handle incoming messages.
-    var message = "" + data;
+    let message = `${data}`;
 
-    message = message.replace(/[\x0D\x0A]/g, "");
+    message = message.replace(/[\x0D\x0A]/g, '');
 
-    if (message.indexOf("/quit") === 0 || data.toString('hex',0,data.length) === "fff4fffd06") {
-      socket.end("Good-bye!\n");
+    if (message.indexOf('/quit') === 0 || data.toString('hex', 0, data.length) === 'fff4fffd06') {
+      socket.end('Good-bye!\n');
       return;
     }
 
     // Use the remoteIP as the name since the PORT changes on ever new connection.
-    bot.reply(socket.remoteAddress, message.trim(), function(err, reply){
-
+    bot.reply(socket.remoteAddress, message.trim(), (err, reply) => {
       // Find the right socket
-      var i = sockets.indexOf(socket);
-      var soc = sockets[i];
+      const i = sockets.indexOf(socket);
+      const soc = sockets[i];
 
-      soc.write("\nBot> " + reply.string + "\n");
-      soc.write("You> ");
-
+      soc.write(`\nBot> ${reply.string}\n`);
+      soc.write('You> ');
     });
   };
 
-  var closeSocket = function(socket, bot) {
-    var i = sockets.indexOf(socket);
-    var soc = sockets[i];
+  const closeSocket = function closeSocket(socket, bot) {
+    const i = sockets.indexOf(socket);
+    const soc = sockets[i];
 
-    console.log("User '" + soc.name + "' has disconnected.\n");
+    console.log(`User '${soc.name}' has disconnected.\n`);
 
-    if (i != -1) {
+    if (i !== -1) {
       sockets.splice(i, 1);
     }
   };
 
-  var newSocket = function (socket) {
-    socket.name = socket.remoteAddress + ":" + socket.remotePort;
-    console.log("User '" + socket.name + "' has connected.\n");
+  const newSocket = function newSocket(socket) {
+    socket.name = `${socket.remoteAddress}:${socket.remotePort}`;
+    console.log(`User '${socket.name}' has connected.\n`);
 
     sockets.push(socket);
-    
+
     // Send a welcome message.
     socket.write('Welcome to the Telnet server!\n');
-    socket.write("Hello " + socket.name + "! " + "Type /quit to disconnect.\n\n");
-
+    socket.write(`Hello ${socket.name}! ` + 'Type /quit to disconnect.\n\n');
 
     // Send their prompt.
-    socket.write("You> ");
+    socket.write('You> ');
 
-    socket.on('data', function(data) {
+    socket.on('data', (data) => {
       receiveData(socket, bot, data);
     });
 
     // Handle disconnects.
-    socket.on('end', function() {
+    socket.on('end', () => {
       closeSocket(socket, bot);
     });
-
   };
 
   // Start the TCP server.
-  var server = net.createServer(newSocket);
+  const server = net.createServer(newSocket);
 
   server.listen(2000);
-  console.log("TCP server running on port 2000.\n");
+  console.log('TCP server running on port 2000.\n');
 };
 
 // This assumes the topics have been compiled to data.json first
-// See superscript/bin/parse for information on how to do that.
+// See superscript/src/bin/parse for information on how to do that.
 
 // Main entry point
-TopicSystem.importerFile('./data.json', function(){
-  new superscript(options, function(err, botInstance){
-    botHandle(null, botInstance);
-  });
+const options = {
+  factSystem: {
+    clean: true,
+  },
+  importFile: './data.json',
+};
+
+superscript.setup(options, (err, bot) => {
+  botHandle(null, bot);
 });
