@@ -1,63 +1,38 @@
-var mocha = require("mocha");
-var should  = require("should");
+import mocha from 'mocha';
+import should from 'should/as-function';
 
-var utils = require("../../lib/utils");
+import utils from '../../src/bot/utils';
 
-describe("Util Helpers", function() {
-
-  it("should not care about sentences with no punctuation", function() {
-    utils.sentenceSplit("Hello world").should.eql([ 'Hello world' ])
+describe('Util Helpers', () => {
+  it('should escape mustaches', () => {
+    should(utils.quotemeta('hello{world}', true)).equal('hello\\{world\\}');
+    should(utils.quotemeta('hello{world}', false)).equal('hello\\{world\\}');
   });
 
-  it("should simple split", function() {
-    utils.sentenceSplit("Hello world.").should.eql([ 'Hello world .' ])
+  it('should only escape pipes when not in commands mode', () => {
+    should(utils.quotemeta('hello|world', true)).equal('hello|world');
+    should(utils.quotemeta('hello|world', false)).equal('hello\\|world');
   });
 
-  it("should double split", function() {
-    utils.sentenceSplit("Hello world. Hello wild world.").should.eql([ 'Hello world .', 'Hello wild world .' ])
+  it('should trim space from string', () => {
+    should(utils.trim('  hello \t\tworld ')).equal('hello world');
   });
 
-  it("should indicate article", function() {
-    utils.indefiniteArticlerize("banana").should.equal("a banana")
-    utils.indefiniteArticlerize("apple").should.equal("an apple")
-    utils.indefiniteArticlerize("hour").should.equal("an hour")
+  it('should preserve newlines in strings', () => {
+    should(utils.trim('  hello \n  world ')).equal('hello \n world');
   });
 
-  it("should indicate article", function() {
-    utils.indefiniteList(["pear", "banana", "apple"]).should.eql("a pear, a banana and an apple")
+  it('should count words', () => {
+    should(utils.wordCount('hello_world#this is a very*odd*string')).equal(8);
   });
 
-  it('should escape mustaches', function() {
-    utils.quotemeta('hello{world}', true).should.equal('hello\\{world\\}')
-    utils.quotemeta('hello{world}', false).should.equal('hello\\{world\\}')
-  })
-
-  it('should only escape pipes when not in commands mode', function() {
-    utils.quotemeta('hello|world', true).should.equal('hello|world')
-    utils.quotemeta('hello|world', false).should.equal('hello\\|world')
-  })
-
-  it('should trim space from string', function() {
-    utils.trim('  hello \t\tworld ').should.equal('hello world')
-  })
-
-  it('should preserve newlines in strings', function() {
-    utils.trim('  hello \n  world ').should.equal('hello \n world')
-  })
-
-  it('should count words', function() {
-    utils.wordCount('hello_world#this is a very*odd*string').should.equal(8)
-  })
-
-  it('should replace captured text', function() {
-    const parts = ['hello <cap>', '', 'how are you <cap2> today', '<cap1>, meet <cap3>']
-    const stars = ['', 'Dave', 'feeling', 'Sally']
-    const replaced = utils.replaceCapturedText(parts, stars)
-    replaced.length.should.equal(3)
-    replaced[0].should.equal('hello Dave')
-    replaced[1].should.equal('how are you feeling today')
-    replaced[2].should.equal('Dave, meet Sally')
-  })
-
+  it('should replace captured text', () => {
+    const parts = ['hello <cap>', '', 'how are you <cap2> today', '<cap1>, meet <cap3>'];
+    const stars = ['', 'Dave', 'feeling', 'Sally'];
+    const replaced = utils.replaceCapturedText(parts, stars);
+    should(replaced.length).equal(3);
+    should(replaced[0]).equal('hello Dave');
+    should(replaced[1]).equal('how are you feeling today');
+    should(replaced[2]).equal('Dave, meet Sally');
+  });
 });
-
