@@ -1,229 +1,20 @@
-var mocha = require("mocha");
-var should  = require("should");
-var help = require("./helpers");
-var async = require("async");
-var Utils = require("../lib/utils");
+/* global describe, it, before, after */
 
-describe('SuperScript Scripting + Style Interface', function(){
-  before(help.before("script"));
+import mocha from 'mocha';
+import should from 'should/as-function';
+import async from 'async';
 
-  describe('Simple star Interface *', function(){
+import helpers from './helpers';
 
-    it("Unscaped", function(done) {
-      bot.reply("user1", "+ this is unscaped", function(err, reply) {
-        reply.string.should.eql("This should pass");
-        done();
-      });
-    });
+describe('SuperScript Scripting + Style Interface', () => {
+  before(helpers.before('script'));
 
-    it("should reply to simple string", function(done) {
-      bot.reply("user1", "This is a test", function(err, reply) {
-        reply.string.should.eql("Test should pass one");
-        done();
-      });
-    });
-
-    it("should match single star", function(done) {
-      bot.reply("user1", "Should match single star", function(err, reply) {
-        ["pass 1", "pass 2", "pass 3"].should.containEql(reply.string);
-        done();
-      });
-    });
-
-    it("should allow empty star - new behaviour", function(done) {
-      bot.reply("user1", "Should match single", function(err, reply) {
-        ["pass 1", "pass 2", "pass 3"].should.containEql(reply.string);
-        done();
-      });
-    });
-
-    it("should match double star", function(done) {
-      bot.reply("user1", "Should match single star two", function(err, reply) {
-        ["pass 1", "pass 2", "pass 3"].should.containEql(reply.string);
-        done();
-      });
-    });
-
-    it("capture in reply", function(done) {
-      bot.reply("user1", "connect the win", function(err, reply) {
-        reply.string.should.eql("Test should pass");
-        done();
-      });
-    });
-
-    it("leading star", function(done) {
-      bot.reply("user1", "my bone", function(err, reply) {
-        reply.string.should.eql("win 1");
-        done();
-      });
-    });
-
-    it("trailing star", function(done) {
-      bot.reply("user1", "bone thug", function(err, reply) {
-        reply.string.should.eql("win 1");
-        done();
-      });
-    });
-
-    it("star star", function(done) {
-      bot.reply("user1", "my bone thug", function(err, reply) {
-        reply.string.should.eql("win 1");
-        done();
-      });
-    });
-
-    it("star star empty", function(done) {
-      bot.reply("user1", "bone", function(err, reply) {
-        reply.string.should.eql("win 1");
-        done();
-      });
-    });
-
-  });
-
-  describe('Exact length star interface *n', function(){
-    it("should match *2 star - Zero case", function(done) {
-      bot.reply("user1", "It is hot out", function(err, reply) {
-        reply.string.should.eql("");
-        done();
-      });
-    });
-
-    it("should match *2 star - One case", function(done) {
-      bot.reply("user1", "It is one hot out", function(err, reply) {
-        reply.string.should.eql("");
-        done();
-      });
-    });
-
-    it("should match *2 star - Two case", function(done) {
-      bot.reply("user1", "It is one two hot out", function(err, reply) {
-        reply.string.should.eql("Test three should pass");
-        done();
-      });
-    });
-
-    it("should match *2 star - Three case", function(done) {
-      bot.reply("user1", "It is one two three hot out", function(err, reply) {
-        reply.string.should.eql("");
-        done();
-      });
-    });
-
-    it("should match *1 star - End case", function(done) {
-      bot.reply("user1", "fixedwidth define love", function(err, reply) {
-        reply.string.should.eql("Test endstar should pass");
-        done();
-      });
-    });
-  });
-
-
-  // min max *(1-2)
-  describe('Mix stars for Mix and Max', function(){
-    it("min max star - Zero", function(done) {
-      bot.reply("user1", "min max", function(err, reply) {
-        reply.string.should.eql("");
-        done();
-      });
-    });
-
-    it("min max star - one", function(done) {
-      bot.reply("user1", "min max one", function(err, reply) {
-        reply.string.should.eql("min max test");
-        done();
-      });
-    });
-
-    it("min max star - two", function(done) {
-      bot.reply("user1", "min max one two", function(err, reply) {
-        reply.string.should.eql("min max test");
-        done();
-      });
-    });
-
-    it("min max star - three", function(done) {
-      bot.reply("user1", "min max one two three", function(err, reply) {
-        reply.string.should.eql("");
-        done();
-      });
-    });
-
-    it("min max star ~emo - gh-221", function(done) {
-      bot.reply("user1", "hello test test", function(err, reply) {
-        reply.string.should.eql("emo reply");
-        done();
-      });
-    });
-
-
-    it.skip("min max star - four", function(done) {
-      bot.reply("user1", "test one. two. three.", function(err, reply) {
-        reply.string.should.eql("test one. two. three.");
-        done();
-      });
-    });
-
-  });
-
-  describe('Variable length star interface *~n', function() {
-    it("should match *~2 star - End case", function(done) {
-      bot.reply("user1", "define love", function(err, reply) {
-        reply.string.should.eql("Test endstar should pass");
-        done();
-      });
-    });
-
-    it("should match *~2 star - Empty", function(done) {
-      bot.reply("user1", "var length", function(err, reply) {
-        ["pass 1"].should.containEql(reply.string);
-        done();
-      });
-    });
-
-
-    it("should match *~2 star - Zero Star", function(done) {
-      bot.reply("user1", "It is hot out 2", function(err, reply) {
-        ["pass 1","pass 2","pass 3"].should.containEql(reply.string);
-        done();
-      });
-    });
-
-    it("should match *~2 star - One Star", function(done) {
-      bot.reply("user1", "It is a hot out 2", function(err, reply) {
-        ["pass 1","pass 2","pass 3"].should.containEql(reply.string);
-        done();
-      });
-    });
-
-    it("should match *~2 star - Two Star", function(done) {
-      bot.reply("user1", "It is a b hot out 2", function(err, reply) {
-        ["pass 1","pass 2","pass 3"].should.containEql(reply.string);
-        done();
-      });
-    });
-
-    it("should match *~2 star - Three Star (fail)", function(done) {
-      bot.reply("user1", "It is a b c d hot out2", function(err, reply) {
-        reply.string.should.eql("");
-        done();
-      });
-    });
-
-    it("should match *~2 star - Return the resuling Star", function(done) {
-      bot.reply("user1", "It is foo bar cold out", function(err, reply) {
-        reply.string.should.eql("Two star result foo bar");
-        done();
-      });
-    });
-  });
-
-  describe('Replies can be repeated accross triggers', function(){
-    it("Replies accross trigger should pass", function(done) {
-      bot.reply("user1", "trigger one", function(err, reply) {
-        reply.string.should.eql("generic reply");
-        bot.reply("user1", "trigger two", function(err, reply) {
-          reply.string.should.eql("generic reply");
+  describe('Replies can be repeated accross triggers', () => {
+    it('Replies accross trigger should pass', (done) => {
+      helpers.getBot().reply('user1', 'trigger 1', (err, reply) => {
+        should(reply.string).eql('generic reply');
+        helpers.getBot().reply('user1', 'trigger 2', (err, reply) => {
+          should(reply.string).eql('generic reply');
           done();
         });
       });
@@ -231,355 +22,188 @@ describe('SuperScript Scripting + Style Interface', function(){
 
     // We exausted this reply in the last test.
     // NB: this test will fail if run on its own.
-    // We lost this functionality when we started walking the tree.
-    it.skip("Should pass 2", function(done) {
-      bot.reply("user1", "trigger one", function(err, reply) {
-        reply.string.should.eql("");
+    it('Should pass 2', (done) => {
+      helpers.getBot().reply('user1', 'trigger one', (err, reply) => {
+        should(reply.string).eql('');
         done();
       });
     });
   });
 
-  describe('Alternates Interface (a|b)', function() {
-    it("should match a or b - Not empty", function(done) {
-      bot.reply("user1", "what is it", function(err, reply) {
-        reply.string.should.eql("");
+  describe('Expand with WordNet', () => {
+    it('should reply to simple string', (done) => {
+      helpers.getBot().reply('user1', 'I love shoes', (err, reply) => {
+        should(reply.string).eql('Wordnet test one');
         done();
       });
     });
 
-    it("should match a or b - should be A", function(done) {
-      bot.reply("user1", "what day is it", function(err, reply) {
-        reply.string.should.eql("Test four should pass");
-        done();
-      });
-    });
-
-    it("should match a or b - should be B", function(done) {
-      bot.reply("user1", "what week is it", function(err, reply) {
-        reply.string.should.eql("Test four should pass");
-        done();
-      });
-    });
-
-    it("should match a or b - word boundries A", function(done) {
-      bot.reply("user1", "what weekend is it", function(err, reply) {
-        reply.string.should.eql("");
-        done();
-      });
-    });
-
-    it("should match a or b - word boundries B", function(done) {
-      bot.reply("user1", "this or that", function(err, reply) {
-        reply.string.should.eql("alter boundry test");
-        done();
-      });
-    });
-
-    it("should match a or b - word boundries C", function(done) {
-      bot.reply("user1", "favorite", function(err, reply) {
-        reply.string.should.eql("");
-        done();
-      });
-    });
-
-    it("should match a or b - word boundries D", function(done) {
-      bot.reply("user1", "this a should e", function(err, reply) {
-        reply.string.should.eql("alter boundry test 2");
-        done();
-      });
-    });
-
-  });
-
-  describe('Optionals Interface [a|b|c]', function(){
-    it("should match empty case", function(done) {
-      bot.reply("user1", "i have a car", function(err, reply) {
-        reply.string.should.eql("Test five should pass");
-        done();
-      });
-    });
-
-    it("should match a", function(done) {
-      bot.reply("user1", "i have a red car", function(err, reply) {
-        reply.string.should.eql("Test five should pass");
-        done();
-      });
-    });
-
-    it("should match b", function(done) {
-      bot.reply("user1", "i have a blue car", function(err, reply) {
-        reply.string.should.eql("Test five should pass");
-        done();
-      });
-    });
-
-    it("should match c", function(done) {
-      bot.reply("user1", "i have a green car", function(err, reply) {
-        reply.string.should.eql("Test five should pass");
-        done();
-      });
-    });
-
-    it("should not match d", function(done) {
-      bot.reply("user1", "i have a black car", function(err, reply) {
-        reply.string.should.eql("");
-        done();
-      });
-    });
-  });
-
-  describe('Expand with WordNet', function() {
-    it("should reply to simple string", function(done) {
-      bot.reply("user1", "I love shoes", function(err, reply) {
-        reply.string.should.eql("Wordnet test one");
-        done();
-      });
-    });
-
-    it("should not expand user-defined concepts greedly (word boundry protection)", function(done) {
-      bot.reply("user1", "I love ballball", function(err, reply) {
-        reply.string.should.eql("");
+    it('should not expand user-defined concepts greedly (word boundry protection)', (done) => {
+      helpers.getBot().reply('user1', 'I love ballball', (err, reply) => {
+        should(reply.string).eql('');
         done();
       });
     });
 
     // This works, but I dont like having to import the DB every time
-    it.skip("should expand user-defined concepts too", function(done) {
-      bot.reply("user1", "I love basketball", function(err, reply) {
-        reply.string.should.eql("Term expanded");
+    it.skip('should expand user-defined concepts too', (done) => {
+      helpers.getBot().reply('user1', 'I love basketball', (err, reply) => {
+        should(reply.string).eql('Term expanded');
         done();
       });
     });
 
     // To match lemma version of wordnet expanded terms, make sure the whole line is lemmed.
-    it.skip("should match both text and lemma", function(done) {
-      bot.reply("user1", "My brother is fat", function(err, reply) {
-        reply.string.should.eql("Ouch");
-        bot.reply("user1", "My brothers is fat", function(err, reply) {
-          reply.string.should.eql("Ouch");
+    it.skip('should match both text and lemma', (done) => {
+      helpers.getBot().reply('user1', 'My brother is fat', (err, reply) => {
+        should(reply.string).eql('Ouch');
+        helpers.getBot().reply('user1', 'My brothers is fat', (err, reply) => {
+          should(reply.string).eql('Ouch');
           done();
         });
-
-      });
-    });
-
-  });
-
-  describe('Replies can have Optionals too!', function(){
-    it("replies with optionals", function(done) {
-      bot.reply("user1", "this reply is random", function(err, reply) {
-        ["yes this reply is awesome","yes this reply is random"].should.containEql(reply.string);
-        done();
-      });
-    });
-
-    it("replies with wordnet", function(done) {
-      bot.reply("user1", "reply with wordnet", function(err, reply) {
-        ["i cotton people","i prefer people", "i care for people", "i love people", "i please people"].should.containEql(reply.string);
-        done();
       });
     });
   });
 
-  describe('Sub-Replies', function(){
-    it("Sub replies 1", function(done) {
-      bot.reply("user1", "redirect_rainbow", function(err, reply) {
+  describe('Replies can have Optionals too!', () => {
+    it('replies with optionals', (done) => {
+      helpers.getBot().reply('user1', 'this reply is random', (err, reply) => {
+        should(['yes this reply is awesome', 'yes this reply is random']).containEql(reply.string);
+        done();
+      });
+    });
 
-        var r = { string: 'red',
+    it('replies with wordnet', (done) => {
+      helpers.getBot().reply('user1', 'reply with wordnet', (err, reply) => {
+        should(['i cotton people', 'i prefer people', 'i care for people', 'i love people', 'i please people']).containEql(reply.string);
+        done();
+      });
+    });
+  });
+
+  describe('Sub-Replies', () => {
+    it('Sub replies 1', (done) => {
+      helpers.getBot().reply('user1', 'redirect_rainbow', (err, reply) => {
+        const r = { string: 'red',
           topicName: 'rainbow',
           subReplies:
-           [ { delay: '500', string: 'orange' },
+          [{ delay: '500', string: 'orange' },
              { delay: '500', string: 'yellow' },
              { delay: '500', string: 'green' },
              { delay: '500', string: 'blue' },
-             { delay: '500', string: 'and black?' } ] };
+             { delay: '500', string: 'and black?' }] };
 
-        reply.should.containDeep(r);
+        should(reply).containDeep(r);
         done();
       });
     });
 
-    it("Sub replies 2", function(done) {
-      bot.reply("user1", "how many colors in the rainbow", function(err, reply) {
-
-        var r = { string: '',
+    it('Sub replies 2', (done) => {
+      helpers.getBot().reply('user1', 'how many colors in the rainbow', (err, reply) => {
+        const r = { string: '',
           topicName: 'rainbow',
           subReplies:
-           [ { delay: '500', string: 'lots' } ] };
+           [{ delay: '500', string: 'lots' }] };
 
-        reply.should.containDeep(r);
+        should(reply).containDeep(r);
         done();
       });
     });
-
-
   });
 
-
-  describe('Custom functions', function(){
-
-    it("should call a custom function with hyphen", function(done) {
-      bot.reply("user1", "error with function thirty-two", function(err, reply) {
-        reply.string.should.eql("thirty-two");
+  describe('Custom functions', () => {
+    it('should call a custom function with hyphen', (done) => {
+      helpers.getBot().reply('user1', 'error with function thirty-two', (err, reply) => {
+        should(reply.string).eql('32');
         done();
       });
     });
 
-
-    it("should call a custom function", function(done) {
-      bot.reply("user1", "custom function", function(err, reply) {
-        reply.string.should.eql("The Definition of function is perform duties attached to a particular office or place or function");
+    it('should call a custom function', (done) => {
+      helpers.getBot().reply('user1', 'custom function', (err, reply) => {
+        should(reply.string).eql('The Definition of function is perform as expected when applied');
         done();
       });
     });
 
-    it("should continue if error is passed into callback", function(done) {
-      bot.reply("user1", "custom 3 function", function(err, reply) {
-        reply.string.should.eql("backup plan");
+    it('should continue if error is passed into callback', (done) => {
+      helpers.getBot().reply('user1', 'custom 3 function', (err, reply) => {
+        should(reply.string).eql('backup plan');
         done();
       });
     });
 
-    it("pass a param into custom function", function(done) {
-      bot.reply("user1", "custom 5 function", function(err, reply) {
-        reply.string.should.eql("he likes this");
+    it('pass a param into custom function', (done) => {
+      helpers.getBot().reply('user1', 'custom 5 function', (err, reply) => {
+        should(reply.string).eql('he likes this');
         done();
       });
     });
 
-    it("pass a param into custom function1", function(done) {
-      bot.reply("user1", "custom 6 function", function(err, reply) {
-        ["he cottons this","he prefers this", "he cares for this", "he loves this", "he pleases this"].should.containEql(reply.string);
+    it('pass a param into custom function1', (done) => {
+      helpers.getBot().reply('user1', 'custom 6 function', (err, reply) => {
+        should(['he cottons this', 'he prefers this', 'he cares for this', 'he loves this', 'he pleases this']).containEql(reply.string);
         done();
       });
     });
 
-    it("the same function twice with different params", function(done) {
-      bot.reply("user1", "custom 8 function", function(err, reply) {
-        reply.string.should.eql("4 + 3 = 7");
+    it('the same function twice with different params', (done) => {
+      helpers.getBot().reply('user1', 'custom 8 function', (err, reply) => {
+        should(reply.string).eql('4 + 3 = 7');
         done();
       });
     });
 
-    it("should not freak out if function does not exist", function(done) {
-      bot.reply("user1", "custom4 function", function(err, reply) {
-        reply.string.should.eql("one + one = 2");
+    it('should not freak out if function does not exist', (done) => {
+      helpers.getBot().reply('user1', 'custom 4 function', (err, reply) => {
+        should(reply.string).eql('one + one = 2');
         done();
       });
     });
 
-    it("function in multi-line reply", function(done) {
-      bot.reply("user1", "custom9 function", function(err, reply) {
-        reply.string.should.eql("a\nb\none\n\nmore");
+    it('function in multi-line reply', (done) => {
+      helpers.getBot().reply('user1', 'custom 9 function', (err, reply) => {
+        should(reply.string).eql('a\nb\none\n\nmore');
         done();
       });
     });
-
   });
-
 
   // I moved this to 5 times because there was a odd chance that we could hit the keep message 2/3rds of the time
-  describe('Reply Flags', function() {
-
-    it("Keep Flag 2", function(done) {
-      bot.reply("user1", "reply flags 2", function(err, reply) {
-        reply.string.should.eql("keep this");
-        bot.reply("user1", "reply flags 2", function(err, reply) {
-          reply.string.should.eql("keep this");
+  describe('Reply Flags', () => {
+    it('Keep Flag 2', (done) => {
+      helpers.getBot().reply('user1', 'reply flags 2', (err, reply) => {
+        should(reply.string).eql('keep this');
+        helpers.getBot().reply('user1', 'reply flags 2', (err, reply) => {
+          should(reply.string).eql('keep this');
           done();
         });
       });
     });
   });
 
-  describe('Custom functions 2 - plugin related', function(){
-    it("Alpha Length 1", function(done) {
-      bot.reply("user1", "How many characters in the word socks?", function(err, reply) {
-        reply.string.should.eql("5");
-        done();
-      });
-    });
-
-    it("Alpha Length 2", function(done) {
-      bot.reply("user1", "How many characters in the name Bill?", function(err, reply) {
-        reply.string.should.eql("4");
-        done();
-      });
-    });
-
-    it("Alpha Length 3", function(done) {
-      bot.reply("user1", "How many characters in the Alphabet?", function(err, reply) {
-        reply.string.should.eql("26");
-        done();
-      });
-    });
-
-    it("Alpha Length 4", function(done) {
-      bot.reply("suser1", "blank", function(err, reply) {
-        bot.getUser("suser1", function(err, u){
-          u.setVar("name", "Bill", function(){
-            bot.reply("suser1", "How many characters in my name?", function(err, reply) {
-              reply.string.should.eql("There are 4 letters in your name.");
-              done();
-            });
-          });
-        });
-      });
-    });
-
-    it("Alpha Lookup 1", function(done) {
-      bot.reply("user1", "What letter comes after B", function(err, reply) {
-        reply.string.should.eql("C");
-        done();
-      });
-    });
-
-    it("Alpha Lookup 2", function(done) {
-      bot.reply("user1", "What letter comes before Z", function(err, reply) {
-        reply.string.should.eql("Y");
-        done();
-      });
-    });
-
-    it("Alpha Lookup 3", function(done) {
-      bot.reply("user1", "What is the last letter in the alphabet?", function(err, reply) {
-        reply.string.should.eql("It is Z.");
-        done();
-      });
-    });
-
-    it("Alpha Lookup 4", function(done) {
-      bot.reply("user1", "What is the first letter of the alphabet?", function(err, reply) {
-        reply.string.should.eql("It is A.");
-        done();
-      });
-    });
-
-  });
-
-  describe('Custom functions 3 - user fact system', function(){
-    it("Should save and recall 1", function(done) {
-      bot.reply("userX", "My name is Bob", function(err, reply) {
-        reply.string.should.eql("Hi Bob.");
-        bot.getUser("userX", function(err, u1){
-          u1.getVar('name', function(err, name){
-            name.should.eql("Bob");
+  describe('Custom functions 3 - user fact system', () => {
+    it('Should save and recall 1', (done) => {
+      helpers.getBot().reply('userX', 'save name Bob', (err, reply) => {
+        should(reply.string).eql('Hi Bob.');
+        helpers.getBot().getUser('userX', (err, u1) => {
+          u1.getVar('name', (err, name) => {
+            should(name).eql('Bob');
             done();
           });
         });
       });
     });
 
-    it("Should save and recall 2", function(done) {
-      bot.reply("suser2", "My name is Ken", function(err, reply) {
-        reply.string.should.eql("Hi Ken.");
-        bot.getUser("userX", function(err, u1){
-          bot.getUser("suser2", function(err, u2){
-            u1.getVar("name", function(err, res){
-              res.should.eql("Bob");
-              u2.getVar("name", function(err, res){
-                res.should.eql("Ken");
+    it('Should save and recall 2', (done) => {
+      helpers.getBot().reply('suser2', 'save name Ken', (err, reply) => {
+        should(reply.string).eql('Hi Ken.');
+        helpers.getBot().getUser('userX', (err, u1) => {
+          helpers.getBot().getUser('suser2', (err, u2) => {
+            u1.getVar('name', (err, res) => {
+              should(res).eql('Bob');
+              u2.getVar('name', (err, res) => {
+                should(res).eql('Ken');
                 done();
               });
             });
@@ -587,25 +211,25 @@ describe('SuperScript Scripting + Style Interface', function(){
         });
       });
     });
-
   });
 
-  describe.skip('Custom functions 4 - user topic change', function(){
-    it("Change topic", function(done) {
-      bot.reply("user3", "call function with new topic", function(err, reply) {
-        bot.reply("user3", "i like fish", function(err, reply) {
-          reply.string.should.eql("me too");
+  describe('Custom functions 4 - user topic change', () => {
+    it('Change topic', (done) => {
+      helpers.getBot().reply('user3', 'call function with new topic', (err, reply) => {
+        helpers.getBot().reply('user3', 'i like fish', (err, reply) => {
+          should(reply.string).eql('me too');
           done();
         });
       });
     });
 
-    it("Change topic 2", function(done) {
-      bot.reply("user4", "reply with a new topic from function", function(err, reply) {
-        bot.getUser("user4", function(err, user){
-          user.currentTopic.should.eql("fish");
-          bot.reply("user4", "i like fish", function(err, reply) {
-            reply.string.should.eql("me too");
+    // This will require processing function tags before any other reply tags
+    it.skip('Change topic 2', (done) => {
+      helpers.getBot().reply('user4', 'reply with a new topic from function', (err, reply) => {
+        helpers.getBot().getUser('user4', (err, user) => {
+          should(user.currentTopic).eql('fish');
+          helpers.getBot().reply('user4', 'i like fish', (err, reply) => {
+            should(reply.string).eql('me too');
             done();
           });
         });
@@ -613,327 +237,412 @@ describe('SuperScript Scripting + Style Interface', function(){
     });
   });
 
-
-  describe('Filter functions', function(){
-    it("Trigger function", function(done) {
-      bot.reply("scuser5", "trigger filter function", function(err, reply) {
-        reply.string.should.eql("");
-        bot.reply("scuser5", "trigger filler function", function(err, reply) {
-          reply.string.should.eql("trigger filter reply");
+  describe('Filter functions', () => {
+    it('Trigger function', (done) => {
+      helpers.getBot().reply('scuser5', 'trigger filter function', (err, reply) => {
+        should(reply.string).eql('');
+        helpers.getBot().reply('scuser5', 'trigger filler function', (err, reply) => {
+          should(reply.string).eql('trigger filter reply');
           done();
         });
       });
     });
   });
 
-  describe('Should parse subfolder', function(){
-    it("Item in folder should exist", function(done) {
-      bot.topicSystem.topic.findOne({name:'suba'}, function(e,res){
-        res.should.not.be.false;
+  describe('Should parse subfolder', () => {
+    it('Item in folder should exist', (done) => {
+      helpers.getBot().chatSystem.Topic.findOne({ name: 'suba' }, (e, res) => {
+        should(res).not.be.false;
         done();
       });
     });
   });
 
-  describe('Emo reply', function(){
-    it("Emo Hello 1", function(done) {
-      bot.reply("user1", "Hello", function(err, reply) {
-        reply.string.should.eql("Hello");
-        done();
-      });
-    });
-  });
+  describe('Filter on Replies', () => {
+    it('should save knowledge', (done) => {
+      helpers.getBot().reply('r1user1', 'my name is Adam.', (err, reply) => {
+        should(reply.string).containEql('Nice to meet you, Adam.');
 
-  describe('Filter on Replies', function(){
-    it("should save knowledge", function(done) {
-      bot.reply("r1user1", "okay my name is Adam.", function(err, reply) {
-        reply.string.should.containEql("Nice to meet you, Adam.");
-        bot.reply("r1user1", "okay my name is Adam.", function(err, reply1) {
-          reply1.string.should.containEql("I know, you already told me your name.");
-          done();
+        // The Reply HAS a filter
+        helpers.getBot().chatSystem.Reply.findOne({ _id: reply.replyId }, (e, res) => {
+          should(res.filter).containEql('^hasName("false")');
+
+          // The user added the fact to the local sublevel
+          helpers.getBot().getUser('r1user1', (err, user) => {
+            user.memory.db.get({ subject: 'name', predicate: 'r1user1' }, (err, results) => {
+              should(results[0].object).containEql('Adam');
+
+              // Now lets hit the other reply / filter
+              helpers.getBot().reply('r1user1', 'my name is Adam.', (err, reply1) => {
+                should(reply1.string).containEql('I know, you already told me your name.');
+                done();
+              });
+            });
+          });
         });
       });
     });
   });
 
-  describe('Augment reply Object', function(){
-    it("Should have replyProp", function(done) {
-      bot.reply("user1", "Can you smile?", function(err, reply) {
-        reply.string.should.eql("Sure can.");
-        reply.emoji.should.eql("smile");
+  describe('Augment reply Object', () => {
+    it('Should have replyProp', (done) => {
+      helpers.getBot().reply('user1', 'Can you smile?', (err, reply) => {
+        should(reply.string).eql('Sure can.');
+        should(reply.emoji).eql('smile');
         done();
       });
     });
 
-    it("Augment callback 1", function(done) {
-      bot.reply("user1", "object param one", function(err, reply) {
-        reply.string.should.eql("world");
-        reply.attachments.should.eql([ { text: 'Optional text that appears *within* the attachment' } ]);
+    it('Augment callback 1', (done) => {
+      helpers.getBot().reply('user1', 'object param one', (err, reply) => {
+        should(reply.string).eql('world');
+        should(reply.attachments).eql([{ text: 'Optional text that appears *within* the attachment' }]);
         done();
       });
     });
 
-    it("Augment callback 2", function(done) {
-      bot.reply("user1", "object param two", function(err, reply) {
-        reply.string.should.eql("world");
-        reply.foo.should.eql("bar");
+    it('Augment callback 2', (done) => {
+      helpers.getBot().reply('user1', 'object param 2', (err, reply) => {
+        should(reply.string).eql('world');
+        should(reply.foo).eql('bar');
         done();
       });
     });
 
     // Params though redirects & Merge
-    it("Augment callback 3", function(done) {
-      bot.reply("user1", "object param three", function(err, reply) {
-        reply.string.should.eql("world");
-        reply.foo.should.eql("bar");
-        reply.attachments.should.eql([ { text: 'Optional text that appears *within* the attachment' } ]);
-        done();
-      });
-    });
-
-  });
-
-  describe('Create Gambit Helper', function(){
-    it("contains concept", function(done) {
-      bot.reply("user1", "my husband likes fish", function(err, reply) {
+    it('Augment callback 3', (done) => {
+      helpers.getBot().reply('user1', 'object param 3', (err, reply) => {
+        should(reply.string).eql('world');
+        should(reply.foo).eql('bar');
+        should(reply.attachments).eql([{ text: 'Optional text that appears *within* the attachment' }]);
         done();
       });
     });
   });
 
-
-  describe('Wrapping lines', function(){
-    it("should continue onto the next line", function(done){
-      bot.reply("user1", "tell me a poem", function(err, reply) {
-        reply.string.should.eql("Little Miss Muffit sat on her tuffet,\nIn a nonchalant sort of way.\nWith her forcefield around her,\nThe Spider, the bounder,\nIs not in the picture today.");
+  describe('Wrapping lines', () => {
+    it('should continue onto the next line', (done) => {
+      helpers.getBot().reply('user1', 'tell me a poem', (err, reply) => {
+        should(reply.string).eql('Little Miss Muffit sat on her tuffet,\nIn a nonchalant sort of way.\nWith her forcefield around her,\nThe Spider, the bounder,\nIs not in the picture today.');
         done();
       });
     });
   });
 
-  describe('Normalize Trigger', function(){
-    it("should be expanded before trying to match", function(done){
-      bot.reply("user1", "it is all good in the hood", function(err, reply) {
-        reply.string.should.eql("normalize trigger test");
+  describe('Normalize Trigger', () => {
+    it('should be expanded before trying to match', (done) => {
+      helpers.getBot().reply('user1', 'it is all good in the hood', (err, reply) => {
+        should(reply.string).eql('normalize trigger test');
         done();
       });
     });
 
-    it("should be expanded before trying to match contract form", function(done){
-      bot.reply("user1", "it's all good in the hood two", function(err, reply) {
-        reply.string.should.eql("normalize trigger test");
+    it('should be expanded before trying to match contract form', (done) => {
+      helpers.getBot().reply('user1', "it's all good in the hood two", (err, reply) => {
+        should(reply.string).eql('normalize trigger test');
         done();
       });
     });
 
-    it("message should exist after normalize", function(done){
-      bot.reply("user1", "then", function(err, reply) {
-        reply.string.should.eql("");
+    it('message should exist after normalize', (done) => {
+      helpers.getBot().reply('user1', 'then', (err, reply) => {
+        should(reply.string).eql('');
         done();
       });
-    });
-
-  });
-
-  describe('Mix case test', function(){
-    it("should match all capitals", function(done){
-      bot.reply("user1", "this is all capitals", function(err, reply) {
-        reply.string.should.eql("Test six should pass");
-        done();
-      });
-    });
-
-    it("should match some capitals", function(done){
-      bot.reply("user1", "this IS ALL capitals", function(err, reply) {
-        reply.string.should.eql("Test six should pass");
-        done();
-      });
-    });
-
-    it("should match with or without puct - 1", function(done){
-      bot.reply("user1", "Do you have a clue?", function(err, reply) {
-        reply.string.should.eql("Test seven should pass");
-        done();
-      });
-    });
-
-    it("should match with or without puct - 2", function(done){
-      bot.reply("user1", "Do you have a cause", function(err, reply) {
-        reply.string.should.eql("Test seven should pass");
-        done();
-      });
-    });
-
-    it("should match with extra spaces mixed in", function(done){
-      bot.reply("user1", "Do       you       have   a    condition", function(err, reply) {
-        reply.string.should.eql("Test seven should pass");
-        done();
-      });
-    });
-
-    it("should allow spaces at the end of replies", function(done){
-      bot.reply("user1", "spaced out", function(err, reply) {
-        reply.string.should.eql("note the space  ");
-        done();
-      });
-    });
-
-  });
-
-  describe('Style - burst related', function(){
-    it("should removed bursted commas", function(done){
-      bot.reply("user1", "John is older than Mary, and Mary is older than Sarah", function(err, reply) {
-        reply.string.should.eql("Test eight should pass");
-        done();
-      });
-    });
-
-    it("should removed bursted commas 2", function(done){
-      bot.reply("user1", "Is it morning, noon, night?", function(err, reply) {
-        reply.string.should.eql("Test nine should pass");
-        done();
-      });
-    });
-
-    it("should removed quotes", function(done){
-      bot.reply("user1", 'remove quotes around "car"?', function(err, reply) {
-        reply.string.should.eql("Test ten should pass");
-        done();
-      });
-    });
-
-    it("should keep reply quotes", function(done){
-      bot.reply("user1", "reply quotes", function(err, reply) {
-        reply.string.should.eql('Test "eleven" should pass');
-        done();
-      });
-    });
-
-    it("dont burst urls", function(done){
-      Utils.sentenceSplit("should not burst http://google.com").should.have.length(1);
-      Utils.sentenceSplit("should not burst 19bdnznUXdHEOlp0Pnp9JY0rug6VuA2R3zK4AACdFzhE").should.have.length(1);
-      Utils.sentenceSplit("burst test should pass rob@silentrob.me").should.have.length(1);
-      done();
     });
   });
 
-  describe('Keep the current topic when a special topic is matched', function(){
-    it("Should redirect to the first gambit", function(done) {
-      bot.reply("user1", "first flow match", function(err, reply) {
-        reply.string.should.eql("You are in the first reply.");
+  describe('Mix case test', () => {
+    it('should match all capitals', (done) => {
+      helpers.getBot().reply('user1', 'this is all capitals', (err, reply) => {
+        should(reply.string).eql('Test six must pass');
+        done();
+      });
+    });
 
-        bot.reply("user1", "second flow match", function(err, reply) {
-          reply.string.should.eql("You are in the second reply. You are in the first reply.");
+    it('should match some capitals', (done) => {
+      helpers.getBot().reply('user1', 'this IS ALL capitals', (err, reply) => {
+        should(reply.string).eql('Test six must pass');
+        done();
+      });
+    });
+
+    it('should match with or without puct - 1', (done) => {
+      helpers.getBot().reply('user1', 'Do you have a clue?', (err, reply) => {
+        should(reply.string).eql('Test seven must pass');
+        done();
+      });
+    });
+
+    it('should match with or without puct - 2', (done) => {
+      helpers.getBot().reply('user1', 'Do you have a cause', (err, reply) => {
+        should(reply.string).eql('Test seven must pass');
+        done();
+      });
+    });
+
+    it('should match with extra spaces mixed in', (done) => {
+      helpers.getBot().reply('user1', 'Do       you       have   a    condition', (err, reply) => {
+        should(reply.string).eql('Test seven must pass');
+        done();
+      });
+    });
+
+    it('should allow spaces at the end of replies', (done) => {
+      helpers.getBot().reply('user1', 'spaced out', (err, reply) => {
+        should(reply.string).eql('note the space  ');
+        done();
+      });
+    });
+  });
+
+  describe('Style - burst related', () => {
+    it('should removed bursted commas', (done) => {
+      helpers.getBot().reply('user1', 'John is older than Mary, and Mary is older than Sarah', (err, reply) => {
+        should(reply.string).eql('Test eight must pass');
+        done();
+      });
+    });
+
+    it('should removed bursted commas 2', (done) => {
+      helpers.getBot().reply('user1', 'Is it morning, noon, night?', (err, reply) => {
+        should(reply.string).eql('Test nine must pass');
+        done();
+      });
+    });
+
+    it('should removed quotes', (done) => {
+      helpers.getBot().reply('user1', 'remove quotes around "car"?', (err, reply) => {
+        should(reply.string).eql('Test ten must pass');
+        done();
+      });
+    });
+
+    it('should keep reply quotes', (done) => {
+      helpers.getBot().reply('user1', 'reply quotes', (err, reply) => {
+        should(reply.string).eql('Test "eleven" must pass');
+        done();
+      });
+    });
+  });
+
+  describe('Keep the current topic when a special topic is matched', () => {
+    it('Should redirect to the first gambit', (done) => {
+      helpers.getBot().reply('user1', 'flow match', (err, reply) => {
+        should(reply.string).eql('You are in the first reply.');
+
+        helpers.getBot().reply('user1', 'next flow match', (err, reply) => {
+          should(reply.string).eql('You are in the second reply. You are in the first reply.');
           done();
         });
       });
     });
 
-    it("Should redirect to the first gambit after matching __pre__", function(done) {
-      bot.reply("user1", "first flow match", function(err, reply) {
-        reply.string.should.eql("You are in the first reply.");
+    it('Should redirect to the first gambit after matching __pre__', (done) => {
+      helpers.getBot().reply('user1', 'flow match', (err, reply) => {
+        should(reply.string).eql('You are in the first reply.');
 
-        bot.reply("user1", "flow redirection test", function(err, reply) {
-          reply.string.should.eql("Going back. You are in the first reply.");
+        helpers.getBot().reply('user1', 'flow redirection test', (err, reply) => {
+          should(reply.string).eql('Going back. You are in the first reply.');
           done();
         });
       });
     });
   });
 
-  describe("gh-173", function(){
-    it("should keep topic though sequence", function(done){
-      bot.reply("user1", "name", function(err, reply) {
-        reply.string.should.eql("What is your first name?");
-        reply.topicName.should.eql("set_name");
+  describe('gh-173', () => {
+    it('should keep topic though sequence', (done) => {
+      helpers.getBot().reply('user1', 'name', (err, reply) => {
+        should(reply.string).eql('What is your first name?');
+        should(reply.topicName).eql('set_name');
 
-        bot.reply("user1", "Bob", function(err, reply) {
-          reply.topicName.should.eql("set_name");
-          reply.string.should.eql("Ok Bob, what is your last name?");
+        helpers.getBot().reply('user1', 'Bob', (err, reply) => {
+          should(reply.topicName).eql('set_name');
+          should(reply.string).eql('Ok Bob, what is your last name?');
 
-          bot.reply("user1", "Hope", function(err, reply) {
+          helpers.getBot().reply('user1', 'Hope', (err, reply) => {
             // this is where we FOUND the reply
-            reply.topicName.should.eql("set_name");
+            should(reply.topicName).eql('set_name');
             // the new topic (pending topic should now be random)
-            bot.getUser("user1", function(err, user){
-              user.getTopic().should.eql("random");
+            helpers.getBot().getUser('user1', (err, user) => {
+              should(user.getTopic()).eql('random');
               done();
             });
           });
-
         });
       });
     });
   });
 
-  describe("scope creep!", function(){
-
-    it("pass scope into redirect", function(done) {
-      bot.reply("user1", "scope though redirect", function(err, reply) {
-        reply.string.should.eql('A user1 __B__');
+  describe('scope creep!', () => {
+    it('pass scope into redirect', (done) => {
+      helpers.getBot().reply('user1', 'scope though redirect', (err, reply) => {
+        should(reply.string).eql('A user1 __B__');
         done();
       }, {
-        key: "A"
+        key: 'A',
       });
     });
 
-    it("dont leak scope", function(done) {
-
+    it('dont leak scope', (done) => {
       async.parallel([
-          function(callback){
-            bot.reply("userA", "generic message", function(err, reply) {
-              callback(null, reply.string);
-            }, {
-              key: "A"
-            });
-
-          },
-          function(callback){
-            bot.reply("userB", "generic message two", function(err, reply) {
-              callback(null, reply.string);
-            }, {
-              key: "B"
-            });
-          }
+        function (callback) {
+          helpers.getBot().reply('userA', 'generic message', (err, reply) => {
+            callback(null, reply.string);
+          }, {
+            key: 'A',
+          });
+        },
+        function (callback) {
+          helpers.getBot().reply('userB', 'generic message two', (err, reply) => {
+            callback(null, reply.string);
+          }, {
+            key: 'B',
+          });
+        },
       ],
       // optional callback
-      function(err, results){
-        results.should.containEql('generic reply A userA generic message');
-        results.should.containEql('generic reply B userB generic message two');
+      (err, results) => {
+        should(results).containEql('generic reply A userA generic message');
+        should(results).containEql('generic reply B userB generic message 2');
 
         done();
       });
     });
   });
 
-  describe('Direct Reply', function() {
-    it("should return reply", function(done) {
-      bot.directReply("user1", "generic", "__simple__", function(err, reply) {
-        reply.string.should.eql("");
+  describe('Direct Reply', () => {
+    it('should return reply', (done) => {
+      helpers.getBot().directReply('user1', 'generic', '__simple__', (err, reply) => {
+        should(reply.string).eql('');
         done();
       });
     });
   });
 
-
-  describe.skip('GH-243', function() {
-    it("Should pass data back into filter function on input", function(done) {
-      bot.reply("user1", "filter by logic", function(err, reply) {
-        reply.string.should.eql("logic");
+  describe('GH-243', () => {
+    it('Should pass data back into filter function on input', (done) => {
+      helpers.getBot().reply('user2', 'filter by logic', (err, reply) => {
+        should(reply.string).eql('logic');
         done();
       });
     });
 
-    it("Should pass data back into filter function on input 2", function(done) {
-      bot.reply("user1", "filter by ai", function(err, reply) {
-        reply.string.should.eql("ai");
+    it('Should pass data back into filter function on input 2', (done) => {
+      helpers.getBot().reply('user2', 'filter by ai', (err, reply) => {
+        should(reply.string).eql('ai');
         done();
       });
     });
-
   });
 
+  describe('GH-301: addMessageProp should work through redirects', () => {
+    it('Should return multiple props', (done) => {
+      helpers.getBot().reply('user2', '__preview', (err, reply) => {
+        should(reply.topLevelProp).exist;
+        should(reply.subProp).exist;
+        done();
+      });
+    });
+  });
 
-  after(help.after);
+  describe('custom functions should work with objects and arrays as parameters', () => {
+    it('Should understand objects and arrays as parameters', (done) => {
+      helpers.getBot().reply('user2', "let's test objects/arrays as custom function args", (err, reply) => {
+        should(reply.string).eql("here's my answer value hey!");
+        done();
+      });
+    });
+  });
 
+  describe('custom functions that return more tags should process them', () => {
+    it('Should process result of custom function', (done) => {
+      helpers.getBot().reply('user2', "what if there's more tags in custom func", (err, reply) => {
+        should(reply.string).eql('and the result is yay');
+        done();
+      });
+    });
+  });
+
+  describe('should use custom tags', () => {
+    it('should respond to different version of saying hello', (done) => {
+      helpers.getBot().reply('user3', 'hi', (err, reply) => {
+        should(reply.string).eql('Greetings!');
+        done();
+      });
+    });
+  });
+
+  describe('gh-265', () => {
+    it('variable length issue simple case', (done) => {
+      helpers.getBot().reply('user5', 'i go by bus', (err, reply) => {
+        should(reply.string).eql('so you go by bus');
+        done();
+      });
+    });
+
+    it('variable length issue fail case', (done) => {
+      helpers.getBot().reply('user5', 'i go by something else', (err, reply) => {
+        should(reply.string).eql('so you go by something else');
+        done();
+      });
+    });
+  });
+
+  describe('gh-312', () => {
+    it('should not crash calling ^createUserFact', (done) => {
+      helpers.getBot().reply('user6', 'set a fact', (err, reply) => {
+        should(reply.string).eql('that is a cool fact');
+        done();
+      });
+    });
+  });
+
+  describe('Simple Question Matching', () => {
+    it('should reply to simple string', (done) => {
+      helpers.getBot().reply('asdf', 'which way to the bathroom?', (err, reply) => {
+        should(reply.string).eql('Down the hall on the left');
+        done();
+      });
+    });
+
+    it('should not match', (done) => {
+      helpers.getBot().reply('asdf', 'My mom cleans the bathroom.', (err, reply) => {
+        should(reply.string).eql('');
+        done();
+      });
+    });
+  });
+
+  describe('gh-237', () => {
+    it('variable length stars should not undercatch', (done) => {
+      helpers.getBot().directReply('user7', 'testfoo', 'foo', (err, reply) => {
+        should(reply.string).eql('Direct match');
+        done();
+      });
+    });
+  });
+
+  describe('gh-171', () => {
+    it('topicRedirects should not skip replies', (done) => {
+      helpers.getBot().reply('user8', 'redirect setup', (err, reply) => {
+        should(reply.string).eql('who are you?');
+        done();
+      });
+    });
+  });
+
+  describe('gh-218', () => {
+    it('replies using redirects should exhaust all the triggers it hits on the way', (done) => {
+      helpers.getBot().reply('user9', 'we should keep this trigger', (err, reply) => {
+        should(reply.string).eql('part one reply some other text i dynamically generate part two reply');
+        helpers.getBot().reply('user9', 'we should keep this trigger', (err, reply) => {
+          should(reply.string).eql('some other text i dynamically generate');
+          done();
+        });
+      });
+    });
+  });
+
+  after(helpers.after);
 });
